{"schema":"decodex/commit/1","summary":"Split core block persistence modules","authority":"manual"}
